+++ server/auth/main.js
@@ -35,7 +35,6 @@ const config  = require('../config')
 
 // -- Local Constants
 const { level } = config
-    , log       = KZlog('auth/main.js', level, false)
     ;
 
 
@@ -58,12 +57,14 @@ const Auth = {
    * @since 0.0.0
    */
   isSession(req, res, next) {
+    const log = KZlog('auth/main.js', level, false);
+
     if (req.session.user_id) {
       // Ok. It's an open session with cookies. Let's go on!
       next();
     } else {
       // It's a request without authentication. Say no!
-      res.status(401).send({ status: 'error', message: 'This request requires an user authentication!' });
+      res.status(401).send({ status: 401, message: 'This request requires an user authentication!' });
       log.warn('This request requires an user authentication!');
     }
   },
@@ -71,25 +72,28 @@ const Auth = {
   /**
    * Connects the authorized clients.
    *
-   * @method (arg1, arg2)
+   * @method (arg1, arg2, arg3)
    * @public
+   * @param {Object}        the database interface object,
    * @param {Object}        Express.js request object,
    * @param {Object}        Express.js response object,
    * @returns {}            -,
    * @since 0.0.0
    */
-  login(req, res) {
+  login(dbi, req, res) {
+    const log = KZlog('auth/main.js', level, false);
+
     log.trace(`user: ${req.body.user}, password: ${req.body.password}.`);
-    Account.getCredentials(req.body.user, req.body.password, (user) => {
+    Account.getCredentials(dbi, req.body.user, req.body.password, (user) => {
       if (user) {
         // Right username and right password:
         req.session.user_id = req.body.user;
-        res.status(200).send({ status: 'Ok', message: 'You are now connected!' });
+        res.status(200).send({ status: 200, message: 'You are now connected!' });
         log.trace(`req.session.user_id: ${req.session.user_id}.`);
         log.trace('You are now connected!');
       } else {
         // Wrong username and/or wrong password:
-        res.status(401).send({ status: 'error', message: `${req.body.user}/${req.body.password} wrong username and/or password!'` });
+        res.status(401).send({ status: 401, message: `${req.body.user}/${req.body.password} wrong username and/or password!'` });
         log.warn(`${req.body.user} + ' doesn't exist or wrong password.`);
       }
     });
@@ -106,8 +110,10 @@ const Auth = {
    * @since 0.0.0
    */
   logout(req, res) {
+    const log = KZlog('auth/main.js', level, false);
+
     delete req.session.user_id;
-    res.status(200).send({ status: 'Ok', message: 'You are now disconnected!' });
+    res.status(200).send({ status: 200, message: 'You are now disconnected!' });
     log.trace('You are now disconnected!');
   },
 };