+++ server/core/routes.js
@@ -33,7 +33,6 @@ const config = require('../config')
 
 // -- Local Constants
 const { level } = config
-    , log       = KZlog('core/routes.js', level, false)
     ;
 
 
@@ -47,13 +46,17 @@ const Routes = {
   /**
    * Starts listening requests from the client web site.
    *
-   * @method (arg1)
+   * @method (arg1, arg2, arg3)
    * @public
    * @param {Object}        the express.js app,
+   * @param {Object}        the message translator,
+   * @param {Object}        the db interface object,
    * @returns {}            -,
    * @since 0.0.0
   */
-  start(app) {
+  start(app, i18n, dbi) {
+    const log = KZlog('core/routes.js', level, false);
+
     // Check if it is an https request:
     app.all('/api/*', (req, res, next) => {
       if (!config.env.https) {
@@ -63,7 +66,7 @@ const Routes = {
         log.info('Kapp is running on Travis-CI, HTTP accesses are exceptionally authorized!');
         next();
       } else if (!req.secure) {
-        res.status(401).send({ status: 'error', message: 'HTTP accesses are not authorized!' });
+        res.status(401).send({ status: 401, message: 'HTTP accesses are not authorized!' });
         log.warn('HTTP accesses are not authorized!');
       } else {
         log.trace('It is an HTTPS request. Authorize it.');
@@ -72,18 +75,18 @@ const Routes = {
     });
 
     // Listen for the implemented api routes:
-    Api.listen(app);
+    Api.listen(app, i18n, dbi);
 
     // Unknown api:
     app.all('/api/*', (req, res) => {
-      res.status(403).send({ status: 'error', message: `The api ${req.method} "${req.url}" does nor exist!` });
-      log.warn(`The api ${req.method} "${req.url}" does nor exist!`);
+      res.status(403).send({ status: 401, message: `${req.method} api "${req.url}" does not exist!` });
+      log.warn(`${req.method} api "${req.url}" does not exist!`);
     });
 
     // Forbidden routes:
     app.all('/*', (req, res) => {
-      res.status(403).send({ status: 'error', message: 'This request is forbidden!' });
-      log.warn('This request is forbidden!');
+      res.status(403).send({ status: 401, message: 'This route is forbidden!' });
+      log.warn('This route is forbidden!');
     });
   },
 };